backup_react/backend/requirements.txt
Flask
flask-cors
python-docx
requests
python-dotenv
boto3 